+++ Tests/Functional/config/routing.yml
@@ -8,6 +8,11 @@ api_root_empty:
     defaults:
         _controller: TestBundle:Root:empty
 
+api_users_proxy_pager:
+    pattern: /api/users/proxy-pager
+    defaults:
+        _controller: TestBundle:User:getUsersProxyPager
+
 api_user_get:
     pattern: /api/users/{identifier}
